requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
motor==3.3.2
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
pymongo==4.6.0
APScheduler==3.10.4
PyJWT==2.8.0